+++ tests/res/dylib/build.rs
@@ -1,0 +1,4 @@
+//! This build script only exists to enable the `OUT_DIR` environment variable
+//! inside the crate.
+
+fn main() {}